モデルの肥大化を防ぐため、メソッドをdecoratorに移動
